pyproject.toml
[tool.poetry]
name = "hostel-waiting-list-checker-backend"
version = "0.1.0"
description = ""
authors = ["Ahmed-Mohiuddin-Shah <mohiuddinlive@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
